create git pull shell block while setting up dbt blocks

+++ ddpui/api/client/prefect_api.py
@@ -15,7 +15,7 @@ from ddpui import auth
 from ddpui.ddpprefect import prefect_service
 from ddpui.ddpairbyte import airbyte_service
 
-from ddpui.ddpprefect import DBTCORE
+from ddpui.ddpprefect import DBTCORE, SHELLOPERATION
 from ddpui.models.org import OrgPrefectBlock, OrgWarehouse, OrgDataFlow
 from ddpui.models.org_user import OrgUser
 from ddpui.ddpprefect.schema import (
@@ -26,6 +26,7 @@ from ddpui.ddpprefect.schema import (
     PrefectDataFlowCreateSchema2,
     PrefectFlowRunSchema,
     PrefectDataFlowUpdateSchema,
+    PrefectShellSetup,
 )
 
 from ddpui.utils.custom_logger import CustomLogger
@@ -357,6 +358,8 @@ def post_prefect_dbt_core_block(request):
     - dbt run
     - dbt test
     - dbt docs generate
+    and Create one shell block to do git pull
+    - git pull
     for a ddp-dbt-profile
     """
     orguser: OrgUser = request.orguser
@@ -378,6 +381,32 @@ def post_prefect_dbt_core_block(request):
 
     if not os.path.exists(dbt_project_filename):
         raise HttpError(400, dbt_project_filename + " is missing")
+
+    # create the git pull shell block
+    try:
+        command = "git pull"
+        block_name = f"{orguser.org.slug}-" f"{slugify(command)}"
+        shell_cmd = PrefectShellSetup(
+            blockname=block_name,
+            commands=[command],
+            workingDir=project_dir,
+            env={},
+        )
+        block_response = prefect_service.create_shell_block(shell_cmd)
+
+        # store prefect shell block in database
+        coreprefectblock = OrgPrefectBlock(
+            org=orguser.org,
+            block_type=SHELLOPERATION,
+            block_id=block_response["block_id"],
+            block_name=block_response["block_name"],
+            display_name=block_name,
+            command=slugify(command),
+        )
+
+    except Exception as error:
+        logger.exception(error)
+        raise HttpError(400, str(error)) from error
 
     with open(dbt_project_filename, "r", encoding="utf-8") as dbt_project_file:
         dbt_project = yaml.safe_load(dbt_project_file)
@@ -496,7 +525,7 @@ def get_prefect_dbt_run_blocks(request):
         }
 
         # fetch the manual deploymentId for the dbt run block
-        if prefect_block.command == "run":
+        if prefect_block.block_type == DBTCORE and prefect_block.command == "run":
             dataflow = OrgDataFlow.objects.filter(
                 org=orguser.org, cron=None, connection_id=None
             ).first()
@@ -519,7 +548,12 @@ def delete_prefect_dbt_run_block(request):
     for dbt_block in org_dbt_blocks:
         # Delete block in prefect
         try:
-            prefect_service.delete_dbt_core_block(dbt_block.block_id)
+            if dbt_block.block_type == DBTCORE:
+                prefect_service.delete_dbt_core_block(dbt_block.block_id)
+
+            if dbt_block.block_type == SHELLOPERATION:
+                prefect_service.delete_shell_block(dbt_block.block_id)
+
         except Exception as error:
             logger.exception(error)
             # may have deleted the block via the prefect ui, continue
